+++ CHANGELOG.md
@@ -3,6 +3,8 @@
     - accepts only to store values at the root level of `d`, and only in `d`.
     - cannot be used in aliases
   - Fix aggregators `aggSum`, `aggAvg`, `aggMin`, `aggMax`, `aggCount` when used with filters (without iterators) like `{d.cars[id=1].wheels[size=100].makers[].qty:aggSum}`
+  - New Formatters `mod()` to compute modulo
+      - Example: `{d.val:mod(2)}` returns 0 if `d.val` = 4
 
 ### v4.5.1
   - Release October 7th 2022